requirements.txt
gemini
langchain
streamlit
python-dotenv
PyPDF2

-e .